+++ roncoo-model/src/main/java/com/model/pojo/Role.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public class Role {
 
-    private Integer id;//主键
+    private long id;//主键
     private Date gmt_create;//创建时间
     private Date gmt_modified;//修改时间
     private Integer status_id;//状态(1:正常，0:禁用)
@@ -12,11 +12,11 @@ public class Role {
     private String role_name;//角色名称
     private String remark;//备注
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
